+++ src/ui/buffers/downloads.rs
@@ -263,6 +263,10 @@ impl Buffer for DownloadsBuffer {
         self
     }
 
+    fn set_theme(&mut self, theme: Theme) {
+        self.theme = theme;
+    }
+
     fn can_close(&self) -> bool {
         true
     }