+++ src/ui/buffers/help.rs
@@ -323,6 +323,10 @@ impl Buffer for HelpBuffer {
         self
     }
 
+    fn set_theme(&mut self, theme: Theme) {
+        self.theme = theme;
+    }
+
     fn can_close(&self) -> bool {
         true
     }