+++ src/ui/buffers/sync.rs
@@ -825,6 +825,10 @@ impl Buffer for SyncBuffer {
         self
     }
 
+    fn set_theme(&mut self, theme: Theme) {
+        self.theme = theme;
+    }
+
     fn can_close(&self) -> bool {
         true
     }